Handle gwt-app packaging the same as war packaging (#1302)
Project which result in a war file are added in their exploded form to
the docker image layers. This is only the case if the packaging is
'war', which poses a problem for GWT projects created by the
gwt-maven-plugin.

This change adds support for the gwt-app packaging in order to use the
jib maven plugin together with the gwt maven plugin.