Optionally write out the image name to a file after build.
If the maven plugin is configured with an `imageName` output
path:
```xml
<configuration>
  <to>
    <image>gcr.io/some/path/to/image:tag</image>
  </to>
  <outputPaths>
    <imageName>target/jib-image.name</imageName>
  </outputPaths>
</configuration>
```

The jib-maven-plugin will now create a file at that path with
the image name during the build, like so:
```
$ mvn jib:build
...
$ cat target/jib-image.name
gcr.io/some/path/to/image:tag
```

This is handy if you generate the image name or tag in a dynamic
fashion (eg, including a timstamp) and you have other tools in a
CD pipeline that operate on the generated image. In our case, we
would like to use the reference to the generated image to deploy to
K8S, and our deployment tooling picks it up from a file.

I have only implemented support in the maven plugin for now, since
that is what we are using and I'm not really familiar with Gradle.
That should be pretty straight-forward to add, if need be.

When it comes to the implementation, I imitated how the image id
and digests are written out for now. Changing to a json or yaml
output format felt like a bigger (breaking) change.

Related issue: #1875