Replace assemble task dependency with runtimeClasspath
This introduces a new lifecycle task called "jibDependencies" that all
the jib tasks depend on and it in turn depends on any requiste tasks
/ configurations. In particular, jibDependencies will depend on the
war/bootWar tasks  when applicable, the jar task when applicable and
always the runtime configuration.  Depending directly on the runtime
configuration allows Gradle to add to the task graph appropriately to
provide all the artifacts that are specified in that configuration,
making it more efficient than using the assemble task.