Fix class cast exception in gradle plugin main class detection logic (#3435)
* Fix class cast exception in gradle plugin main class detection logic

The main class detection logic in the gradle plugin always assumed that
the `Main-Class` attribute, set in the main jar task, is a string. This
is not always true, as it sometimes makes sense to set it to a provider
of String in order to link multiple configurations together.

The logic now tests the class of the value returned for the `Main-Class`
attribute and in case of an `Provider<?>` value, it will call the `get`
method. If the provider is not set, this will cause the task to fail.
The resulting object is then transformed to a String.

fixes #3396

* Update jib-gradle-plugin/CHANGELOG.md

Co-authored-by: Chanseok Oh <chanseok@google.com>