Prototype having platform-dependent layers
Despite building Java containers, it's not uncommon for platform-specific dependencies to be needed. Common examples include custom JREs produced via jlink and native libraries used via JNI. The image for each platform will typically still have the same layers, but the contents of one of the layers will differ.

To model this, a new "platform layer" is introduced. It associates various platforms with the file entries to be used for that platform. Importantly, the layer still shows as a single layer in the overall build plan, since each image must contain 1 such layer. To achieve this, a new LayerObject, the PlatformDependentLayer, is added to the API. The jib-core project is updated to understand this new LayerObject and select the platform-dependent entry based on the platform of the base image.